Move store into separate module
