--- src/app/shared/ngrx/instruments/instruments.actions.ts
@@ -1,4 +1,0 @@
-import { createAction, props } from '@ngrx/store';
-import { Instrument } from '../../models/instruments/instrument.model';
-
-export const selectInstrument = createAction('[Instruments] SelectInstrument', props<{ instrument: Instrument }>());
