--- src/app/shared/ngrx/portfolios/portfolios.actions.ts
@@ -1,8 +1,0 @@
-import { createAction, props } from '@ngrx/store';
-import { PortfolioKey } from "../../models/portfolio-key.model";
-
-export const selectPortfolio = createAction('[Portfolios] SelectPortfolio', props<{ portfolio: PortfolioKey | null }>());
-
-
-
-
