+++ src/app/store/instruments/instruments.reducer.ts
@@ -1,6 +1,6 @@
 import { createReducer, on } from '@ngrx/store';
 import * as InstrumentsActions from './instruments.actions';
-import { Instrument } from "../../models/instruments/instrument.model";
+import { Instrument } from '../../shared/models/instruments/instrument.model';
 
 export const instrumentsFeatureKey = 'instruments';
 
@@ -20,7 +20,7 @@ export const initialState: InstrumentsState = {
 export const reducer = createReducer(
   initialState,
 
-  on(InstrumentsActions.selectInstrument, (state, { instrument }) => ({
+  on(InstrumentsActions.selectNewInstrument, (state, { instrument }) => ({
       ...state,
       selectedInstrument: {
         ...instrument
