+++ src/app/store/portfolios/portfolios.reducer.ts
@@ -1,6 +1,6 @@
 import { createReducer, on } from '@ngrx/store';
-import { PortfolioKey } from '../../models/portfolio-key.model';
 import * as PortfoliosActions from './portfolios.actions';
+import { PortfolioKey } from '../../shared/models/portfolio-key.model';
 
 export const portfoliosFeatureKey = 'portfolios';
 
@@ -15,7 +15,7 @@ export const initialState: PortfoliosState = {
 export const reducer = createReducer(
   initialState,
 
-  on(PortfoliosActions.selectPortfolio, (state, { portfolio }) => ({
+  on(PortfoliosActions.selectNewPortfolio, (state, { portfolio }) => ({
       ...state,
       selectedPortfolio: portfolio
     })
